move utils fields and constants
