+++ misc_crypto/poseidon/__init__.py
@@ -1,6 +1,5 @@
 from .poseidon import (
-    get_pseudo_random,
-    get_matrix,
     Poseidon,
     poseidon_t6,
 )
+from .utils import get_pseudo_random, get_matrix
