+++ misc_crypto/poseidon/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_SEED = b"poseidon"
