+++ misc_crypto/poseidon/poseidon.py
@@ -1,51 +1,8 @@
-from py_ecc.bn128 import curve_order
-from py_ecc.fields import FQ
-from hashlib import blake2b
 from eth_utils import to_tuple
 from typing import Tuple, Iterator, Sequence
-
-
-# All the algebra of the circuit must be in the Fr Field
-class Fr(FQ):
-    field_modulus = curve_order
-
-
-@to_tuple
-def get_pseudo_random(seed: bytes, n: int) -> Iterator[Fr]:
-    h = seed
-    for _ in range(n):
-        h = blake2b(h, digest_size=32).digest()
-        yield Fr(int.from_bytes(h, "little"))
-
-
-def all_different(_tuple: Sequence[Fr]) -> bool:
-    _list = list(_tuple)
-    for _ in range(len(_tuple)):
-        a = _list.pop()
-        if a == Fr(0):
-            return False
-        for b in _list:
-            if a == b:
-                return False
-    return True
-
-
-@to_tuple
-def get_matrix(t: int, seed: bytes) -> Iterator[Tuple[Fr, ...]]:
-    nonce = 0
-    _seed = b"".join([seed, b"_matrix_", f"{nonce:04d}".encode()])
-    cmatrix = get_pseudo_random(_seed, t * t)
-    while not all_different(cmatrix):
-        nonce += 1
-        nonceStr = f"{nonce:04d}".encode() if nonce < 10000 else bytes(nonce)
-        _seed = b"".join([seed, b"_matrix_", nonceStr])
-        cmatrix = get_pseudo_random(_seed, t * t)
-    for i in range(t):
-        yield tuple(Fr(1) / (cmatrix[i] - cmatrix[t + j]) for j in range(t))
-
-
-def get_constants(t: int, seed: bytes, rounds: int):
-    return get_pseudo_random(seed + b"_constants", rounds)
+from .constants import DEFAULT_SEED
+from .fields import Fr
+from .utils import get_constants, get_matrix, recommend_parameter
 
 
 def ark(state: Sequence[Fr], c: Fr) -> Tuple[Fr, ...]:
@@ -73,31 +30,6 @@ def mix(state: Tuple[Fr], M: Tuple[Tuple[Fr], ...]) -> Iterator[Fr]:
         yield accumulation
 
 
-def recommend_parameter(elements_length: int) -> Tuple[int, int, int]:
-    """
-    A faster way to get recommended parameters from the result of running parameter_finder.py.
-    """
-    if elements_length < 2:
-        raise ValueError("Number of elements should be at least 2")
-
-    # Poseidon requires 1 width to be reserved for security.
-    t = elements_length + 1
-    roundsF = 8
-    if t == 3:
-        roundsP = 49
-    elif t <= 7:
-        roundsP = 50
-    elif t <= 14:
-        roundsP = 51
-    elif t <= 29:
-        roundsP = 52
-    else:
-        raise ValueError(
-            "Unsupported elements_length, please run parameter_finder.py to get the parameter"
-        )
-    return t, roundsF, roundsP
-
-
 class Poseidon:
     seed: bytes
     roundsF: int
@@ -106,7 +38,7 @@ class Poseidon:
     matrix: Tuple[Tuple[Fr, ...], ...]
     constants: Tuple[Fr]
 
-    def __init__(self, t: int, roundsF: int, roundsP: int, seed=b"poseidon") -> None:
+    def __init__(self, t: int, roundsF: int, roundsP: int, seed=DEFAULT_SEED) -> None:
         self.t = t
         self.roundsF = roundsF
         self.roundsP = roundsP
